Перечень исправлений: -Устранено задвоение наименований .bcfv и .png файлов при потворном сохранении BCF-файла; -Для соответствия требованиям формата BCF реализовано создание пустого viewpoint.bcfv файла для каждого Topic-а при сериализации; -Реализована возможность работы с несколькими файлами .bcfv в одном Topic-е; -Изменен конструктор класса Topic, теперь в конструкторе создается пустой список List<KeyValuePair<String, VisualizationXMLFile>> Visualizations (по аналогии со Snapshots); -Исправлены типы данных для аргументов масштаб/угол обзора в конструкторах камер. -В сеттере свойства ID класса BCFViewpoint убрана проверка на Guid.Empty.

+++ Xbim.BCF/XMLNodes/BCFOrthogonalCamera.cs
@@ -86,7 +86,7 @@ namespace Xbim.BCF.XMLNodes
         private BCFOrthogonalCamera()
         { }
 
-        public BCFOrthogonalCamera(Vector cameraViewPoint, Vector cameraDirection, Vector cameraUpVector, int viewToWorldScale)
+        public BCFOrthogonalCamera(Vector cameraViewPoint, Vector cameraDirection, Vector cameraUpVector, double viewToWorldScale)
         {
             CameraViewPoint = cameraViewPoint;
             CameraDirection = cameraDirection;
